third_party/blink/web_tests/platform/linux-chrome/external/wpt/reporting/document-reporting-destroy-after-document-close.https.sub-expected.txt

This is a testharness.js-based test.
[FAIL] No more reports received after navigation to same document without endpoint header
  assert_equals: expected 1 but got 0
Harness: the test ran to completion.
